CASMINST-7394 Add workflowRef label unique to image to avoid overwriting
We use workflowRef label to identify workflow, which needs to be re-triggered
to update image with latest CVE fixes. We also try to re-trigger base image,
but workflowRef label of base image is susually overwritten by descendant image build.
To avoid that, we will add workflowRef.<image> label, which is unique to
image name.